+++ lopatina.sofia/T2/stream_guard.cpp
@@ -1,0 +1,14 @@
+#include "stream_guard.hpp"
+
+lopatina::StreamGuard::StreamGuard(std::basic_ios<char> & s):
+  s_(s),
+  fill_(s.fill()),
+  precision_(s.precision()),
+  flags_(s.flags())
+ {}
+lopatina::StreamGuard::~StreamGuard()
+{
+  s_.fill(fill_);
+  s_.precision(precision_);
+  s_.flags(flags_);
+}
